% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_classes.R
\name{as_percentage}
\alias{as_percentage}
\alias{as_currency}
\alias{as_hyperlink}
\alias{as_scientific}
\title{Assign openxlsx format to column}
\usage{
as_percentage(x)

as_currency(x)

as_hyperlink(x)

as_scientific(x)
}
\arguments{
\item{x}{Column}
}
\description{
Explicit openxlsx columnar class assignment.
}
\section{Functions}{
\itemize{
\item \code{as_percentage}: Assign percentage format to column

\item \code{as_currency}: Assign currency format to column

\item \code{as_hyperlink}: Assign hyperlink format to column

\item \code{as_scientific}: Assign scientific format to column
}}

\examples{
library(magrittr)
tab <-
 data.frame(
   type = c("a", "b"),
   abs = c(60, 40),
   share = c(.6, .4)
 )
tab$share <- as_percentage(tab$share)

# Column have includes percentage class and will be proper exported to excel sheet:
class(tab$share)
}
