% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsheet.R
\name{sheet_dim}
\alias{sheet_dim}
\title{Get sheet's last row and column for the data which is already placed on this sheet.}
\usage{
sheet_dim(workbook, sheet = 1)
}
\arguments{
\item{workbook}{Workbook object}

\item{sheet}{Sheet name or index}
}
\description{
This function can be used in case you need add smth to the current workbook.
}
