% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlbox-package.R
\docType{package}
\name{xlbox-package}
\alias{xlbox}
\alias{xlbox-package}
\title{xlbox: Simple export of standard R objects to xlsx}
\description{
Provide fast, simple and customizable xlsx export on top of `openxlsx` package, based on simple nested list structure.
}
\author{
\strong{Maintainer}: Vladimir Korovin \email{vladimir.korovin@gfk.com}

}
\keyword{internal}
